requirements.txt
sqlalchemy==2.0.49
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pytest==8.3.3